fix MSVS build

+++ SDL_crossfire/game/widgets/Bonuses.hpp
@@ -35,7 +35,7 @@ private:
 	static constexpr uint8_t ENABLE_NUMB = 12;
 	static constexpr uint8_t DISABLE_NUMB = 6;
 	static constexpr double ANIMATION_PERIOD = 500;
-	static constexpr double TEXT_TO = 2000;
+	static constexpr uint32_t TEXT_TO = 2000;
 
 	struct Position_t {
 		Point m_EnablePos;
